Close #7: Dont show all templates
We add a configuration option to actually whitelist the choosable
themes (aka. templates). This appears to be more general than some
special naming convention to blacklist themes, and also takes away the
need to make a decision in this regard.

For best backward compatibility, and because this is supposedly the
best default, we choose `*` (i.e. all templates).